Issue verification code on mailing list sign up
Extract the logic to issue a verification code into a concern that can be
reused in both the event and mailing list sign up.
