+++ spec/requests/mailing_list/steps_controller_spec.rb
@@ -2,6 +2,7 @@ require "rails_helper"
 
 describe MailingList::StepsController do
   include_context "stub types api"
+  include_context "stub candidate create access token api"
 
   let(:model) { MailingList::Steps::Name }
   let(:step_path) { mailing_list_step_path model.key }
