+++ spec/support/wizard_support.rb
@@ -15,6 +15,49 @@ shared_examples "a wizard step" do
   it { is_expected.to respond_to :save }
 end
 
+shared_examples "an issue verification code wizard step" do
+  describe "#save" do
+    before do
+      subject.email = "email@address.com"
+      subject.first_name = "first"
+      subject.last_name = "last"
+    end
+
+    let(:request) do
+      GetIntoTeachingApiClient::ExistingCandidateRequest.new(
+        email: subject.email,
+        firstName: subject.first_name,
+        lastName: subject.last_name,
+      )
+    end
+
+    context "when invalid" do
+      it "does not call the API" do
+        subject.email = nil
+        subject.save
+        expect_any_instance_of(GetIntoTeachingApiClient::CandidatesApi).to_not receive(:create_candidate_access_token)
+      end
+    end
+
+    context "when an existing candidate" do
+      it "sends verification code and sets authenticate to true" do
+        allow_any_instance_of(GetIntoTeachingApiClient::CandidatesApi).to receive(:create_candidate_access_token).with(request)
+        subject.save
+        expect(wizardstore["authenticate"]).to be_truthy
+      end
+    end
+
+    context "when a new candidate or CRM is unavailable" do
+      it "will skip the authenticate step" do
+        allow_any_instance_of(GetIntoTeachingApiClient::CandidatesApi).to receive(:create_candidate_access_token).with(request)
+          .and_raise(GetIntoTeachingApiClient::ApiError)
+        subject.save
+        expect(wizardstore["authenticate"]).to be_falsy
+      end
+    end
+  end
+end
+
 class TestWizard < Wizard::Base
   class Name < Wizard::Step
     attribute :name
